Refactor
- move class from src/access to src/app
- remove config flags
