+++ src/access/BUILD.gn
@@ -23,8 +23,6 @@ static_library("access") {
     "AuthMode.h",
     "Privilege.h",
     "RequestPath.h",
-    "RequiredPrivilege.cpp",
-    "RequiredPrivilege.h",
     "SubjectDescriptor.h",
     "examples/ExampleAccessControlDelegate.cpp",
     "examples/ExampleAccessControlDelegate.h",
