+++ src/app/CommandHandler.cpp
@@ -24,10 +24,10 @@
 
 #include "CommandHandler.h"
 #include "InteractionModelEngine.h"
+#include "RequiredPrivilege.h"
 #include "messaging/ExchangeContext.h"
 
 #include <access/AccessControl.h>
-#include <access/RequiredPrivilege.h>
 #include <app/util/MatterCallbacks.h>
 #include <credentials/GroupDataProvider.h>
 #include <lib/support/TypeTraits.h>
@@ -259,7 +259,7 @@ CHIP_ERROR CommandHandler::ProcessCommandDataIB(CommandDataIB::Parser & aCommand
         Access::SubjectDescriptor subjectDescriptor = mpExchangeCtx->GetSessionHandle()->GetSubjectDescriptor();
         Access::RequestPath requestPath{ .cluster = concretePath.mClusterId, .endpoint = concretePath.mEndpointId };
         Access::Privilege requestPrivilege =
-            Access::RequiredPrivilege::ForInvokeCommand(concretePath.mClusterId, concretePath.mEndpointId, concretePath.mCommandId);
+            RequiredPrivilege::ForInvokeCommand(concretePath.mClusterId, concretePath.mEndpointId, concretePath.mCommandId);
         err = Access::GetAccessControl().Check(subjectDescriptor, requestPath, requestPrivilege);
         err = CHIP_NO_ERROR; // TODO: remove override
         if (err != CHIP_NO_ERROR)
