+++ src/app/util/ember-compatibility-functions.cpp
@@ -22,10 +22,10 @@
  */
 
 #include <access/AccessControl.h>
-#include <access/RequiredPrivilege.h>
 #include <app/ClusterInfo.h>
 #include <app/ConcreteAttributePath.h>
 #include <app/InteractionModelEngine.h>
+#include <app/RequiredPrivilege.h>
 #include <app/reporting/Engine.h>
 #include <app/reporting/reporting.h>
 #include <app/util/af.h>
@@ -388,7 +388,7 @@ CHIP_ERROR ReadSingleClusterData(const SubjectDescriptor & aSubjectDescriptor, b
     {
         Access::RequestPath requestPath{ .cluster = aPath.mClusterId, .endpoint = aPath.mEndpointId };
         Access::Privilege requestPrivilege =
-            Access::RequiredPrivilege::ForReadAttribute(aPath.mClusterId, aPath.mEndpointId, aPath.mAttributeId);
+            RequiredPrivilege::ForReadAttribute(aPath.mClusterId, aPath.mEndpointId, aPath.mAttributeId);
         CHIP_ERROR err = Access::GetAccessControl().Check(aSubjectDescriptor, requestPath, requestPrivilege);
         err            = CHIP_NO_ERROR; // TODO: remove override
         if (err != CHIP_NO_ERROR)
@@ -821,7 +821,7 @@ CHIP_ERROR WriteSingleClusterData(const SubjectDescriptor & aSubjectDescriptor, 
     {
         Access::RequestPath requestPath{ .cluster = aPath.mClusterId, .endpoint = aPath.mEndpointId };
         Access::Privilege requestPrivilege =
-            Access::RequiredPrivilege::ForWriteAttribute(aPath.mClusterId, aPath.mEndpointId, aPath.mAttributeId);
+            RequiredPrivilege::ForWriteAttribute(aPath.mClusterId, aPath.mEndpointId, aPath.mAttributeId);
         CHIP_ERROR err = Access::GetAccessControl().Check(aSubjectDescriptor, requestPath, requestPrivilege);
         err            = CHIP_NO_ERROR; // TODO: remove override
         if (err != CHIP_NO_ERROR)
